fixe#937
